requirements.txt
flask
opencv-python
easyocr
numpy
tesseract-ocr
gunicorn
opencv-python
torch
torchvision
Pillow